chat/pyproject.toml
[project]
name = "auran-chat"
version = "0.1.0"
description = "Auran Chat Server — chat.auran.llc"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.34.0",
    "psycopg2-binary>=2.9.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
]

[project.scripts]
auran-chat = "server:main"
